requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
transformers>=4.35.0
torch>=2.0.0
torchaudio>=2.0.0
pyannote.audio==3.1.1
ffmpeg-python==0.2.0
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
jinja2==3.1.2
python-dotenv==1.0.0
accelerate>=0.20.0
datasets>=2.14.0
